+++ Wiki/ReleaseNotes.md
@@ -2,6 +2,7 @@
 
 ## 7.8.2
 
+- Linq driver for MongoDB now default to version 3.
 - Better support for cancellation token. 
 
 ## 7.8.1